+++ lib/widget/order/so_line_detail.dart
@@ -3,20 +3,23 @@
 /*
  * Widget for displaying detail view of a single SalesOrderLineItem
  */
+
 import "package:flutter/material.dart";
 import "package:flutter_speed_dial/flutter_speed_dial.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
 
-import "package:inventree/app_colors.dart";
-import "package:inventree/l10.dart";
 import "package:inventree/inventree/part.dart";
 import "package:inventree/inventree/sales_order.dart";
+
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/widget/progress.dart";
 import "package:inventree/widget/part/part_detail.dart";
-
-import "package:inventree/helpers.dart";
 import "package:inventree/widget/snacks.dart";
+
+import "package:inventree/app_colors.dart";
+import "package:inventree/l10.dart";
+import "package:inventree/helpers.dart";
+import "package:inventree/api_form.dart";
 
 
 class SoLineDetailWidget extends StatefulWidget {
@@ -34,6 +37,8 @@ class SoLineDetailWidget extends StatefulWidget {
 class _SOLineDetailWidgetState extends RefreshableState<SoLineDetailWidget> {
 
   _SOLineDetailWidgetState();
+
+  InvenTreeSalesOrder? order;
 
   @override
   String getAppBarTitle() => L10().lineItem;
@@ -53,6 +58,53 @@ class _SOLineDetailWidgetState extends RefreshableState<SoLineDetailWidget> {
     }
 
     return actions;
+  }
+
+  Future<void> _allocateStock(BuildContext context) async {
+
+    if (order == null) {
+      return;
+    }
+
+    Map<String, dynamic> fields = {
+      "line_item": {
+        "parent": "items",
+        "nested": true,
+        "hidden": true,
+        "value": widget.item.pk,
+      },
+      "stock_item": {
+        "parent": "items",
+        "nested": true,
+        "filters": {
+          "part": widget.item.partId,
+          "in_stock": true,
+        }
+      },
+      "quantity": {
+        "parent": "items",
+        "nested": true,
+        "value": widget.item.unallocatedQuantity,
+      },
+      "shipment": {
+        "filters": {
+          "order": order!.pk.toString(),
+        }
+      },
+    };
+
+    launchApiForm(
+      context,
+      L10().allocateStock,
+      order!.allocate_url,
+      fields,
+      method: "POST",
+      icon: FontAwesomeIcons.rightToBracket,
+      onSuccess: (data) async {
+        refresh(context);
+      }
+    );
+
   }
 
   Future<void> _editLineItem(BuildContext context) async {
@@ -76,13 +128,35 @@ class _SOLineDetailWidgetState extends RefreshableState<SoLineDetailWidget> {
 
   @override
   List<SpeedDialChild> actionButtons(BuildContext context) {
-    // TODO
-    return [];
+
+    List<SpeedDialChild> buttons = [];
+
+    if (order != null && order!.isOpen) {
+      buttons.add(
+        SpeedDialChild(
+          child: FaIcon(FontAwesomeIcons.rightToBracket, color: Colors.blue),
+          label: L10().allocateStock,
+          onTap: () async {
+            _allocateStock(context);
+          }
+        )
+      );
+    }
+
+    return buttons;
   }
 
   @override
   Future<void> request(BuildContext context) async {
     await widget.item.reload();
+
+    final so = await InvenTreeSalesOrder().get(widget.item.orderId);
+
+    if (mounted) {
+      setState(() {
+        order = (so is InvenTreeSalesOrder ? so : null);
+      });
+    }
   }
 
   @override
@@ -105,6 +179,30 @@ class _SOLineDetailWidgetState extends RefreshableState<SoLineDetailWidget> {
             Navigator.push(context, MaterialPageRoute(builder: (context) => PartDetailWidget(part)));
           }
         }
+      )
+    );
+
+    // Available quantity
+    tiles.add(
+      ListTile(
+        title: Text(L10().availableStock),
+        leading: FaIcon(FontAwesomeIcons.boxesStacked),
+        trailing: Text(simpleNumberString(widget.item.availableStock))
+      )
+    );
+
+    // Allocated quantity
+    tiles.add(
+      ListTile(
+        leading: FaIcon(FontAwesomeIcons.clipboardCheck),
+        title: Text(L10().allocated),
+        subtitle: ProgressBar(widget.item.allocatedRatio),
+        trailing: Text(
+          widget.item.allocatedString,
+          style: TextStyle(
+            color: widget.item.isAllocated ? COLOR_SUCCESS : COLOR_WARNING
+          )
+        )
       )
     );